fixed the broken code
